fix(client): change codebase to typescript
