+++ client/src/components/Editor/Editor.css
@@ -9,8 +9,8 @@
     justify-content: center;
 }
 
-.uploaded-image{
+.uploaded-image {
     margin-top: 20px;
     max-height: 65%;
     max-width: 65%;
-} 
+}
