+++ client/src/components/ImageUploader/ImageUploader.tsx
@@ -5,11 +5,11 @@ import './ImageUploader.css';
 
 const ImageUploader = () => {
   const [image, setImage] = useState(null);
-  const [imageUrl, setImageUrl] = useState('');
+  const [imageUrl, setImageUrl] = useState<string | ArrayBuffer | null>('');
 
   const handleChange = (event) => {
     const { target: { files } } = event;
-    let reader = new FileReader();
+    let reader: FileReader = new FileReader();
     reader.readAsDataURL(files[0]);
     reader.onload = () => {
       setImageUrl(reader.result);
@@ -20,18 +20,18 @@ const ImageUploader = () => {
   const removeImage = () => {
     setImage(null);
   };
-  
+
   return (
     <div>
-      <Header showPublish={image} handleDataAction={removeImage}/>
+      <Header showPublish={Boolean(image)} handleDataAction={removeImage} />
       {!image ?
         <div className="background-container">
           <label htmlFor="files" className="button-position">Upload</label>
           <input id="files" className="input-file" type="file" accept="image/*" onChange={handleChange} />
         </div>
         : <div className="background-container">
-            <Editor imgData={image} imgUrl={imageUrl}/>
-          </div>
+          <Editor imgData={image} imgUrl={imageUrl} />
+        </div>
       }
     </div>
   )
