+++ client/src/components/Toolbar/Toolbar.stories.tsx
