+++ client/src/components/common/Button.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+type Props = {
+	buttonName: string,
+	className?: string
+};
+
+const Button = ({ buttonName, className }: Props) => (
+	<input type="button" name={buttonName} value={buttonName} className={className} />
+);
+
+Button.defaultProps = {
+	className: ''
+};
+
+export default Button;
